feat: add field and form states
